Merge get_path and get_model into one function
obs_data_get_string returns a const char * pointing to the real string.
obs_module_file returns a newly allocated char * which should be freed.
There is no way for the code to know which one was used when freeing
filter->modelname. So the merged version of these functions bstrdups
the string internally only when necessary. The returned string should
then always be freed with bfree. There is no need to cast as they are
already char *.