Leak fixes (#6)
* Don't try to free modelname or maskctx if they are null

* Merge get_path and get_model into one function

obs_data_get_string returns a const char * pointing to the real string.
obs_module_file returns a newly allocated char * which should be freed.
There is no way for the code to know which one was used when freeing
filter->modelname. So the merged version of these functions bstrdups
the string internally only when necessary. The returned string should
then always be freed with bfree. There is no need to cast as they are
already char *.

* Don't free the filter if creation failed

If creating maskctx failed it could be due to bad configuration. In
that case we still need to create the filter in order for the user to
fix the config. We just don't start the thread, so return early, the
same as in update.

* Don't try to strcmp nullptr

Either filter->modelname or model can be nullptr here. If they both
point to the same string, or both null, do nothing. Only strcmp them
if neither is nullptr. Re-init if exactly one is nullptr or if strcmp
says they are different.

* Do not try to stop the thread if it is not running

Doing so would raise a std::system_error, which crashes OBS.

* Free temporary model name if it equals what is already in the filter

If we didn't make changes because the old and new model names strcmp
equal, we need to free the new one.